# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/requests:
    get:
      operationId: getStatsRequests
      tags: [stats]
      summary: Get recent requests
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Recent requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsRequestsResponse"
...
components:
  schemas:
...
    StatsRequest:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        clientKey:
          type: string
        provider:
          type: string
        model:
          type: string
        promptTokens:
          type: integer
          nullable: true
        completionTokens:
          type: integer
          nullable: true
        totalTokens:
          type: integer
          nullable: true
        latencyMs:
          type: integer
          nullable: true
        costUsd:
          type: number
          nullable: true
        status:
          type: integer
        isStream:
          type: boolean
          nullable: true
        requestPath:
          type: string
        errorMessage:
          type: string
          nullable: true
      required:
        - id
        - createdAt
        - clientKey
        - provider
        - model
        - status
        - requestPath
    StatsRequestsResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/StatsRequest"
        total:
          type: integer
      required:
        - requests
        - total
...