# lib/api-zod/src/generated/types/statsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsRequest {
  id: number;
  createdAt: string;
  clientKey: string;
  provider: string;
  model: string;
  promptTokens?: number | null;
  completionTokens?: number | null;
  totalTokens?: number | null;
  latencyMs?: number | null;
  costUsd?: number | null;
  status: number;
  isStream?: boolean | null;
  requestPath: string;
  errorMessage?: string | null;
}