# lib/api-zod/src/generated/types/statsRequestsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatsRequest } from "./statsRequest";

export interface StatsRequestsResponse {
  requests: StatsRequest[];
  total: number;
}